+++ lib/rubocop/cop/doctolib_cops.rb
@@ -1,1 +1,2 @@
 # frozen_string_literal: true
+require_relative 'doctolib/no_asymmetrical_pundit_after_actions'