+++ test/rubocop/doctolib_test.rb
@@ -2,12 +2,8 @@
 
 require "test_helper"
 
-class Rubocop::DoctolibTest < Minitest::Test
+class RuboCop::DoctolibTest < Minitest::Test
   def test_that_it_has_a_version_number
-    refute_nil ::Rubocop::Doctolib::VERSION
-  end
-
-  def test_it_does_something_useful
-    assert false
+    refute_nil ::RuboCop::Doctolib::VERSION
   end
 end