+++ CHANGELOG.md
@@ -1,5 +1,4 @@
-## [Unreleased]
-
 ## [0.1.0] - 2021-07-20
 
-- Initial release
+- Initial release.
+- Add the Doctolib/NoAsymmetricalPunditAfterActions cop.